This commit includes: 1. Removal of test_process_attestation.py from Capella 2. test_validator_appears_only_once_in_attestation has been updated to work with all phases and moved to phase0. Assertion simplified 3. Updates to test_validator_invalid_high_funding_after_withdrawal to handle electra deposits and to correctly handle processing of the deposit through run_deposit_processing
