+++ tests/core/pyspec/eth2spec/test/capella/block_processing/test_process_deposit.py
@@ -62,9 +62,10 @@ def test_success_top_up_to_withdrawn_validator(spec, state):
 @spec_state_test
 def test_validator_invalid_high_funding_after_withdrawal(spec, state):
     """
-    Simple extension of the previous test_success_top_up_to_withdrawn_validator in order to add a negative scenario
-    where a user deposits 2x MAX_EFFECTIVE_BALANCE.
-    """
+     Simple extension of the previous test_success_top_up_to_withdrawn_validator in order to add a negative scenario
+     where a user deposits 2x MAX_EFFECTIVE_BALANCE.
+     Have added support for the increased Electra deposit amount
+     """
     validator_index = 5  # Example validator index
 
     # Fully withdraw the validator
@@ -78,9 +79,18 @@ def test_validator_invalid_high_funding_after_withdrawal(spec, state):
     next_epoch_via_block(spec, state)
     assert state.validators[validator_index].effective_balance == 0
 
-    # Make a top-up balance to validator - attempt to deposit double the max deposit)
-    deposit = spec.MAX_EFFECTIVE_BALANCE * 2
+    # Perform a check to see if the spec is Electra version
+    if is_post_electra(spec):
+        max_effective_balance = spec.MAX_EFFECTIVE_BALANCE_ELECTRA
+    else:
+        max_effective_balance = spec.MAX_EFFECTIVE_BALANCE
 
-    # Handle the negative test based on the readme - no post state
-    if deposit > spec.MAX_EFFECTIVE_BALANCE:
-        yield 'post', None
+    # Make a top-up deposit, attempting to deposit double the max effective balance
+    amount = max_effective_balance * 2
+    deposit = prepare_state_and_deposit(spec, state, validator_index, amount, signed=True)
+
+    # Process the invalid deposit, setting valid=False as we expect it to be rejected
+    run_deposit_processing(spec, state, validator_index, deposit, valid=False)
+
+    # Validate that the deposit was rejected by checking the effective balance
+    assert state.validators[validator_index].effective_balance == 0
